Switch sequence receiver to using the same hash table mech

+++ common/svc_message.cpp
@@ -1201,7 +1201,7 @@ odaproto::svc::MidPrint SVC_MidPrint(const std::string& message, const int time)
 	return msg;
 }
 
-odaproto::svc::ServerGametic SVC_ServerGametic(const byte tic, const byte queueDepth)
+odaproto::svc::ServerGametic SVC_ServerGametic(const byte tic, const int queueDepth)
 {
 	odaproto::svc::ServerGametic msg;
